chore: added erc20gated test
