+++ contracts/paymasters/ERC20gatedPaymaster.sol
@@ -12,7 +12,7 @@ import "@matterlabs/zksync-contracts/l2/system-contracts/Constants.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
 
 /// @author Matter Labs
-/// @notice This smart contract pays the gas fees for accounts with balance of a specific ERC20 token.
+/// @notice This smart contract pays the gas fees for accounts with balance of a specific ERC20 token. It makes use of the approval-based flow paymaster.
 contract ERC20gatedPaymaster is IPaymaster, Ownable {
     uint256 constant PRICE_FOR_PAYING_FEES = 1;
 
@@ -32,7 +32,7 @@ contract ERC20gatedPaymaster is IPaymaster, Ownable {
     }
 
     function validateAndPayForPaymasterTransaction(
-        bytes32,
+        bytes32,    
         bytes32,
         Transaction calldata _transaction
     )
