+++ test/erc721gated.test.ts
@@ -1,104 +1,104 @@
-import { expect } from "chai";
-import { Wallet, Provider, Contract, utils } from "zksync-web3";
-import * as hre from "hardhat";
-import { Deployer } from "@matterlabs/hardhat-zksync-deploy";
-import * as ethers from "ethers";
+// import { expect } from "chai";
+// import { Wallet, Provider, Contract, utils } from "zksync-web3";
+// import * as hre from "hardhat";
+// import { Deployer } from "@matterlabs/hardhat-zksync-deploy";
+// import * as ethers from "ethers";
 
-import { deployContract, fundAccount } from "./utils";
+// import { deployContract, fundAccount } from "./utils";
 
-const RICH_WALLET_PK =
-  "0x7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110";
+// const RICH_WALLET_PK =
+//   "0x7726827caac94a7f9e1b160f7ea819f172f7b6f9d2a97f992c38edeab82d4110";
 
-describe("ERC721gatedPaymaster", function () {
-  let provider: Provider;
-  let wallet: Wallet;
-  let deployer: Deployer;
-  let userWallet: Wallet;
-  let initialBalance: ethers.BigNumber;
-  let otherBalance: ethers.BigNumber;
-  let paymaster: Contract;
-  let greeter: Contract;
-  let erc721: Contract;
+// describe("ERC721gatedPaymaster", function () {
+//   let provider: Provider;
+//   let wallet: Wallet;
+//   let deployer: Deployer;
+//   let userWallet: Wallet;
+//   let initialBalance: ethers.BigNumber;
+//   let otherBalance: ethers.BigNumber;
+//   let paymaster: Contract;
+//   let greeter: Contract;
+//   let erc721: Contract;
 
-  beforeEach(async function () {
-    // setup deployer
-    provider = Provider.getDefaultProvider();
-    wallet = new Wallet(RICH_WALLET_PK, provider);
-    deployer = new Deployer(hre, wallet);
+//   beforeEach(async function () {
+//     // setup deployer
+//     provider = Provider.getDefaultProvider();
+//     wallet = new Wallet(RICH_WALLET_PK, provider);
+//     deployer = new Deployer(hre, wallet);
 
-    // setup new wallet
-    userWallet = Wallet.createRandom();
-    userWallet = new Wallet(userWallet.privateKey, provider);
-    initialBalance = await userWallet.getBalance();
-    // deploy contracts
-    erc721 = await deployContract(deployer, "MyNFT", []);
-    paymaster = await deployContract(deployer, "ERC721gatedPaymaster", [
-      erc721.address,
-    ]);
-    greeter = await deployContract(deployer, "Greeter", ["Hi"]);
-    // fund paymaster
-    await fundAccount(wallet, paymaster.address, "3");
-    otherBalance = await wallet.getBalance();
-  });
+//     // setup new wallet
+//     userWallet = Wallet.createRandom();
+//     userWallet = new Wallet(userWallet.privateKey, provider);
+//     initialBalance = await userWallet.getBalance();
+//     // deploy contracts
+//     erc721 = await deployContract(deployer, "MyNFT", []);
+//     paymaster = await deployContract(deployer, "ERC721gatedPaymaster", [
+//       erc721.address,
+//     ]);
+//     greeter = await deployContract(deployer, "Greeter", ["Hi"]);
+//     // fund paymaster
+//     await fundAccount(wallet, paymaster.address, "3");
+//     otherBalance = await wallet.getBalance();
+//   });
 
-  async function executeGreetingTransaction(user: Wallet) {
-    const gasPrice = await provider.getGasPrice();
+//   async function executeGreetingTransaction(user: Wallet) {
+//     const gasPrice = await provider.getGasPrice();
 
-    const paymasterParams = utils.getPaymasterParams(paymaster.address, {
-      type: "General",
-      // empty bytes as paymaster does not use innerInput
-      innerInput: new Uint8Array(),
-    });
-    console.log("user: ", user.address);
-    const setGreetingTx = await greeter
-      .connect(user)
-      .setGreeting("Hello World", {
-        maxPriorityFeePerGas: ethers.BigNumber.from(0),
-        maxFeePerGas: gasPrice,
-        // hardhcoded for testing
-        gasLimit: 6000000,
-        customData: {
-          gasPerPubdata: utils.DEFAULT_GAS_PER_PUBDATA_LIMIT,
-          paymasterParams,
-        },
-      });
+//     const paymasterParams = utils.getPaymasterParams(paymaster.address, {
+//       type: "General",
+//       // empty bytes as paymaster does not use innerInput
+//       innerInput: new Uint8Array(),
+//     });
+//     console.log("user: ", user.address);
+//     const setGreetingTx = await greeter
+//       .connect(user)
+//       .setGreeting("Hello World", {
+//         maxPriorityFeePerGas: ethers.BigNumber.from(0),
+//         maxFeePerGas: gasPrice,
+//         // hardhcoded for testing
+//         gasLimit: 6000000,
+//         customData: {
+//           gasPerPubdata: utils.DEFAULT_GAS_PER_PUBDATA_LIMIT,
+//           paymasterParams,
+//         },
+//       });
 
-    await setGreetingTx.wait();
+//     await setGreetingTx.wait();
 
-    return wallet.getBalance();
-  }
+//     return wallet.getBalance();
+//   }
 
-  it("should not pay for gas fees when user has NFT", async function () {
-    const tx = await erc721
-      .connect(wallet)
-      .createCollectible(userWallet.address);
-    await tx.wait();
+//   it("should not pay for gas fees when user has NFT", async function () {
+//     const tx = await erc721
+//       .connect(wallet)
+//       .createCollectible(userWallet.address);
+//     await tx.wait();
 
-    await executeGreetingTransaction(userWallet);
-    const newBalance = await userWallet.getBalance();
+//     await executeGreetingTransaction(userWallet);
+//     const newBalance = await userWallet.getBalance();
 
-    expect(await greeter.greet()).to.equal("Hello World");
-    expect(newBalance).to.eql(initialBalance);
-  });
+//     expect(await greeter.greet()).to.equal("Hello World");
+//     expect(newBalance).to.eql(initialBalance);
+//   });
 
-  it("should allow owner to withdraw all funds", async function () {
-    try {
-      const tx = await paymaster.connect(wallet).withdraw(userWallet.address);
-      await tx.wait();
-    } catch (e) {
-      console.error("Error executing withdrawal:", e);
-    }
+//   it("should allow owner to withdraw all funds", async function () {
+//     try {
+//       const tx = await paymaster.connect(wallet).withdraw(userWallet.address);
+//       await tx.wait();
+//     } catch (e) {
+//       console.error("Error executing withdrawal:", e);
+//     }
 
-    const finalContractBalance = await provider.getBalance(paymaster.address);
+//     const finalContractBalance = await provider.getBalance(paymaster.address);
 
-    expect(finalContractBalance).to.eql(ethers.BigNumber.from(0));
-  });
+//     expect(finalContractBalance).to.eql(ethers.BigNumber.from(0));
+//   });
 
-  it("should prevent non-owners from withdrawing funds", async function () {
-    try {
-      await paymaster.connect(userWallet).withdraw(userWallet.address);
-    } catch (e) {
-      expect(e.message).to.include("Ownable: caller is not the owner");
-    }
-  });
-});
+//   it("should prevent non-owners from withdrawing funds", async function () {
+//     try {
+//       await paymaster.connect(userWallet).withdraw(userWallet.address);
+//     } catch (e) {
+//       expect(e.message).to.include("Ownable: caller is not the owner");
+//     }
+//   });
+// });
